+++ src/java/pl/mn/communicator/Status.java
@@ -23,7 +23,7 @@ import java.util.Date;
  * Created on 2004-12-12
  * 
  * @author <a href="mailto:mati@sz.home.pl">Mateusz Szczap</a>
- * @version $Id: Status.java,v 1.6 2004-12-20 22:42:41 winnetou25 Exp $
+ * @version $Id: Status.java,v 1.7 2004-12-21 20:03:29 winnetou25 Exp $
  */
 public class Status implements IStatus {
 
@@ -41,7 +41,7 @@ public class Status implements IStatus {
 	private int m_remotePort = 1555;
 	
 	private int m_imageSize = -1;
-	private byte m_version = -1;
+	private int m_version = -1;
 	private int m_descriptionSize = 0;
 
 	private boolean m_supportsVoiceCommunication = false;
@@ -177,11 +177,11 @@ public class Status implements IStatus {
 	/**
 	 * @see pl.mn.communicator.IStatus#getGGVersion()
 	 */
-	public byte getGGVersion() {
+	public int getGGVersion() {
 		return m_version;
 	}
 	
-	public void setGGVersion(byte version) {
+	public void setGGVersion(int version) {
 		m_version = version;
 	}
